[Refactor]: 작성, 수정 페이지 컴포넌트화 및 통합
